+++ CHANGES.rst
@@ -10,8 +10,8 @@ Internals
 * ``apply_N`` was introduced in module ``mathics.builtin.numeric`` was used to speed up critically used built-in function ``N``. Its use reduces the use of
   ``Expression("N", expr, prec).evaluate(evaluation)`` which is slower.
 * A bug was fixed relating to the order in which ``mathics.core.definitions`` stores the rules
+* Improving support for working with ``Series``.
 * ``Cylinder`` rendering is implemented in Asymptote
-
 
 4.0.1
 -----
@@ -22,6 +22,7 @@ New builtins
 * ``Guidermannian``
 * ``Cone``
 * ``Tube``
+* ``Normal`` now have a basic support for ``SeriesData``
 
 Tensor functions: